Refactor into some kind of sanity
