+++ osu.Game/Overlays/Changelog/ChangelogHeader.cs
@@ -27,8 +27,8 @@ namespace osu.Game.Overlays.Changelog
 
         public ChangelogHeader()
         {
-            TabControl.AddItem(listing_string);
-            TabControl.Current.ValueChanged += e =>
+            BreadcrumbControl.AddItem(listing_string);
+            BreadcrumbControl.Current.ValueChanged += e =>
             {
                 if (e.NewValue == listing_string)
                     ListingSelected?.Invoke();
@@ -46,7 +46,7 @@ namespace osu.Game.Overlays.Changelog
         [BackgroundDependencyLoader]
         private void load(OsuColour colours)
         {
-            TabControl.AccentColour = colours.Violet;
+            BreadcrumbControl.AccentColour = colours.Violet;
             TitleBackgroundColour = colours.GreyVioletDarker;
             ControlBackgroundColour = colours.GreyVioletDark;
         }
@@ -56,12 +56,12 @@ namespace osu.Game.Overlays.Changelog
         private void showBuild(ValueChangedEvent<APIChangelogBuild> e)
         {
             if (e.OldValue != null)
-                TabControl.RemoveItem(e.OldValue.ToString());
+                BreadcrumbControl.RemoveItem(e.OldValue.ToString());
 
             if (e.NewValue != null)
             {
-                TabControl.AddItem(e.NewValue.ToString());
-                TabControl.Current.Value = e.NewValue.ToString();
+                BreadcrumbControl.AddItem(e.NewValue.ToString());
+                BreadcrumbControl.Current.Value = e.NewValue.ToString();
 
                 Streams.Current.Value = Streams.Items.FirstOrDefault(s => s.Name == e.NewValue.UpdateStream.Name);
 
@@ -69,7 +69,7 @@ namespace osu.Game.Overlays.Changelog
             }
             else
             {
-                TabControl.Current.Value = listing_string;
+                BreadcrumbControl.Current.Value = listing_string;
                 Streams.Current.Value = null;
                 title.Version = null;
             }
