+++ osu.Game/Overlays/News/NewsHeader.cs
@@ -24,9 +24,9 @@ namespace osu.Game.Overlays.News
 
         public NewsHeader()
         {
-            TabControl.AddItem(front_page_string);
+            BreadcrumbControl.AddItem(front_page_string);
 
-            TabControl.Current.ValueChanged += e =>
+            BreadcrumbControl.Current.ValueChanged += e =>
             {
                 if (e.NewValue == front_page_string)
                     ShowFrontPage?.Invoke();
@@ -38,7 +38,7 @@ namespace osu.Game.Overlays.News
         [BackgroundDependencyLoader]
         private void load(OsuColour colours)
         {
-            TabControl.AccentColour = colours.Violet;
+            BreadcrumbControl.AccentColour = colours.Violet;
             TitleBackgroundColour = colours.GreyVioletDarker;
             ControlBackgroundColour = colours.GreyVioletDark;
         }
@@ -46,18 +46,18 @@ namespace osu.Game.Overlays.News
         private void showPost(ValueChangedEvent<string> e)
         {
             if (e.OldValue != null)
-                TabControl.RemoveItem(e.OldValue);
+                BreadcrumbControl.RemoveItem(e.OldValue);
 
             if (e.NewValue != null)
             {
-                TabControl.AddItem(e.NewValue);
-                TabControl.Current.Value = e.NewValue;
+                BreadcrumbControl.AddItem(e.NewValue);
+                BreadcrumbControl.Current.Value = e.NewValue;
 
                 title.IsReadingPost = true;
             }
             else
             {
-                TabControl.Current.Value = front_page_string;
+                BreadcrumbControl.Current.Value = front_page_string;
                 title.IsReadingPost = false;
             }
         }
