+++ osu.Game/Overlays/OverlayHeader.cs
@@ -84,7 +84,7 @@ namespace osu.Game.Overlays
                                 RelativeSizeAxes = Axes.Both,
                                 Colour = Color4.Gray,
                             },
-                            CreateControl().With(control => control.Margin = new MarginPadding { Left = UserProfileOverlay.CONTENT_X_MARGIN })
+                            CreateTabControl().With(control => control.Margin = new MarginPadding { Left = UserProfileOverlay.CONTENT_X_MARGIN })
                         }
                     },
                     CreateContent()
@@ -99,6 +99,6 @@ namespace osu.Game.Overlays
 
         protected abstract ScreenTitle CreateTitle();
 
-        protected abstract TabControl<string> CreateControl();
+        protected abstract TabControl<string> CreateTabControl();
     }
 }
